pyproject.toml
[tool.poetry]
name = "soundklips"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
Flask = "^2.0.2"
pandas = "^1.3.3"
gunicorn = "^20.1.0"
PyJWT = "^2.3.0"
psycopg2-binary = "^2.9.1"
pytest = "^6.2.5"
Flask-Cors = "^3.0.10"
Flask-SQLAlchemy = "^2.5.1"
SQLAlchemy = "^1.4.26"
Flask-Mail = "^0.9.1"
google-cloud-storage = "^1.43.0"
librosa = "^0.8.1"
PySoundFile = "^0.9.0"
audioread = "^2.1.9"
Flask-RESTful = "^0.3.9"

[tool.poetry.dev-dependencies]
Flask = "^2.0.2"
pandas = "^1.3.3"
psycopg2-binary = "^2.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
